refactor: 유틸성 클래스에 private 생성자 추가
